web-backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
httpx==0.28.1
a2a-sdk==0.3.22

